Fix the viewmodel tests
