+++ app/src/test/java/com/androidpositive/ghubstore/presentation/repositorylist/list/RepositoryListViewModelTest.kt
@@ -1,14 +1,12 @@
 package com.androidpositive.ghubstore.presentation.repositorylist.list
 
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import com.androidpositive.extensions.MainCoroutineScopeRule
 import com.androidpositive.extensions.createCaptor
 import com.androidpositive.ghubstore.data.interactors.GithubRepositoryInteractor
+import com.androidpositive.viewmodel.BaseViewModelTest
 import com.androidpositive.viewmodel.Resource
 import com.androidpositive.viewmodel.toResource
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
-import org.junit.Rule
 import org.junit.Test
 import org.kohsuke.github.GHRepository
 import org.mockito.kotlin.any
@@ -16,15 +14,7 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 
 @ExperimentalCoroutinesApi
-class RepositoryListViewModelTest {
-    @Rule
-    @JvmField
-    val rule = InstantTaskExecutorRule()
-
-    @Rule
-    @JvmField
-    val coroutineScope = MainCoroutineScopeRule()
-
+class RepositoryListViewModelTest: BaseViewModelTest() {
     private val interactor = mock<GithubRepositoryInteractor>()
 
     @Test
@@ -32,7 +22,7 @@ class RepositoryListViewModelTest {
         val repositories = Result.success(listOf<GHRepository>())
         whenever(interactor.getRepositories(any())).thenReturn(repositories)
 
-        val viewModel = RepositoryListViewModel(interactor)
+        val viewModel = RepositoryListViewModelImpl(interactor)
         val repositoriesCaptor = viewModel.repositories.createCaptor()
 
         repositoriesCaptor.assertSendsValues(
@@ -46,7 +36,7 @@ class RepositoryListViewModelTest {
         val error = Result.failure<List<GHRepository>>(Throwable())
         whenever(interactor.getRepositories(any())).thenReturn(error)
 
-        val viewModel = RepositoryListViewModel(interactor)
+        val viewModel = RepositoryListViewModelImpl(interactor)
         val repositoriesCaptor = viewModel.repositories.createCaptor()
 
         repositoriesCaptor.assertSendsValues(
