+++ infrastructure/test/livedata/build.gradle
@@ -12,5 +12,5 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4'
     implementation 'junit:junit:4.13.2'
-    testImplementation "androidx.arch.core:core-testing:2.1.0"
+    implementation "androidx.arch.core:core-testing:2.1.0"
 }
